+++ backend/helpus_donation/helpus/src/main/java/com/ssafy/helpus/donation/service/Impl/DonationServiceImpl.java
@@ -97,7 +97,7 @@ public class DonationServiceImpl implements DonationService {
 
     @Override
     @Transactional
-    public Map<String, Object> getDonation(Integer donationId) throws Exception {
+    public Map<String, Object> getDonation(Long donationId) throws Exception {
         log.info("DonationService getDonation call");
 
         Map<String, Object> resultMap = new HashMap<>();
@@ -124,7 +124,7 @@ public class DonationServiceImpl implements DonationService {
     }
 
     @Override
-    public Map<String, Object> endDonation(Integer donationId) throws Exception {
+    public Map<String, Object> endDonation(Long donationId) throws Exception {
         log.info("DonationService endDonation call");
 
         Map<String, Object> resultMap = new HashMap<>();
@@ -142,7 +142,7 @@ public class DonationServiceImpl implements DonationService {
     }
 
     @Override
-    public Map<String, Object> listDonation(Integer memberId, String order, int page) {
+    public Map<String, Object> listDonation(Long memberId, String order, int page) {
         log.info("DonationService listDonation call");
 
         Map<String, Object> resultMap = new HashMap<>();
